### 2019-10-02 #345 @SD1060
presubmission: SDesti
Submitting Author: Diogo Sayanda (@SD1060)
Repository:   https://github.com/SD1060/SDesti

---

```
Package: SDesti
Title: An R package for the analysis of macroinvertebrate and periphyton environmental studies’ data
Version: 0.1.0
Date:2019-03-27
Author: Diogo Sayanda
Maintainer: Diogo Sayanda <dsayanda@yahoo.com>
Description: An adequate and complete statistical analysis for Stone dwelling organisms and Periphyton biomonitoring data. It receives data on macroinvertebrates (counts or biological index), or Periphyton (counts, chlorophyll concentration, ash free dry mass (AFDM), or biological index), environmental variables and site habitat composition data collected over a period of time and has four user accessible functions that return: 1) the descriptive analysis of each sample and a surficial evaluation of the quality of the data recurring to the power function; 2) estimates of number of individuals/chlorophyll concentration/AFDM/index per area of benthic habitat at one given site; 3) the typical descriptive analysis that precedes the adjustment of a regression model, including detection of multi-collinearity and determining the transformation that maximizes the linear relation between the response and the covariates in analysis; 4) the adjustment of a regression model (linear, Poisson, quasipoisson, or negative binomial) according to the characteristics of the data, selection of the best model recurring to automated backward and forward routines, estimation of measures of goodness-of-fit, identification of the outliers, influent observations, and delivery of the graphics necessary to analyze the residuals. The program delivers information that allows the user to make a general evaluation of the quality of each sample, evaluate the quality of the analysis and interpret the data. The program includes a complete manual.
License: BSD3 + file LICENSE
Copyright: Diogo Sayanda and Alberta Enviroment and Parks
Imports: grDevices, broom, car, caret, faraway, graphics, MASS, stats
Suggests: knitr,
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

The program is intended to be used in multiple environmental monitoring programs increasing their transparency, consistency and facilitating the integration of distinct studies on surface aquatic systems management.

Biologists that monitor the heath of rivers' aquatic ecosystems.

There is no other package that simplify so much the data analysis as SDesti does.
To perform the exact same analysis the user could recur to numerous functions within other packages which means that the process would be much more complex and time consuming than it is when using SDesti. The selection of the "best" quasipoisson model and testing multiple interaction terms would also have to be done manually. SDesti includes algorithms that perform these automatically while respecting the hierarchical principle.

Chlorophyll data tables are usually complex and chlorophyll concentration, or AFDM have to be calculated. SDesti performs all the needed calculations regarding to this data and reduces the complexity of the analysis and human error.

SDesti was developed for non statisticians and persons that are not proficient with R, delivering the information that they need to interpret the data and reducing time for data analysis and human error.

I don't think so...
