+++ README.md
@@ -1,0 +1,94 @@
+---
+title: "SDesti"
+author: "Diogo Sayanda"
+date: "June, 2019"
+output:
+  pdf_document:
+    toc: yes
+    toc_depth: '6'
+  word_document:
+    toc: yes
+    toc_depth: '6'
+  html_document:
+    highlight: pygments
+    keep_md: yes
+    theme: united
+    toc: yes
+    toc_depth: 6
+vignette: >
+  %\VignetteIndexEntry{"SDesti"} 
+  %\VignetteEncoding{UTF-8}
+  %\VignetteEngine{knitr::knitr}
+---
+
+# Copy rights
+
+This program was developed by Diogo Sayanda and is property of Alberta Government, Canada. Any use, appropriation or distribution of part or totality of this program, manual or any of its components must be formally authorized by the Alberta Government, Canada.
+
+# Description
+
+This package has been designed primarily to be used with biomonitoring epilithic organisms, Chlorophyll and Ash Free Dry Mass data from aquatic systems but can also be used to analyze any epibenthic or endobenthic organisms? data.
+
+Epilithic organisms are any organisms that occur, on stony surfaces, or crevices, and can be quantified (number of organisms per stone, or any community related index), these can be: macroinvertebrates, periphyton, or macrophytes.
+
+Periphyton Chlorophyll content and Ash Free Dry Mass data are extracted from samples collected by scraping the surface of stones, that are posteriorly treated following traditional laboratory protocols (Dowbush, et al. 2017).
+
+The referred endpoint variables are identified by the program through the acronyms $SD$, $Index$, $Chla$ and 
+$AFDM$:
+
+*	$SD$ ? The original expression behind this acronym is ?Stone-Dwelling organisms?. In practice ?SD? indicates to the program that the data refers to number of individuals per stone, or sample replicate, and that the data file has the structure of the SData file (see below). This acronym can be used when the response is represented by the number of individuals (any taxon, or group of taxa) per replicate;
+
+*	$Index$ ? This acronym indicates the program that the response variable is any biological index that resumes the community data for each replicate, and that the data follows exactly the IData file structure;
+
+*	$Chla$ ? The original expression for this acronym is ?Chlorophyll-a? and it is the information that the program needs, to know that the data structure of the file complies exactly with the Chla data file (see below), and that the response is the quantity of any pigment, or metabolite, that was taken from biofilm scraped from rocky surfaces;
+
+*	$AFDM$ ? This acronym is the most restrictive of all, refers to ?Ash Free Dry Mass? of periphyton samples, and it is the information the program needs, to know that the data has the structure of the Chla data file and that it must proceed with the needed calculations of AFDM.
+
+Besides biological data, it is possible to include any environmental data related with physical, chemical, or biotic parameters to analyze the correlation between these covariates and the responses. The corresponding file must comply with the structure of the EData file (see below).
+
+The present package is simple to use and performs the complete macroinvertebrate, or periphyton, data analysis, delivering the necessary information to evaluate the quality of the data, to analyze the quality of the results and adequacy of the analysis, to interpret the results and to answer to the objectives of environmental studies.
+
+The program contains four user assessable functions that perform the following tasks:
+
+#.	Data quality evaluation and descriptive analysis of a $Chla$, $AFDM$, $Index$, or $SD$ single data sample;
+
+#.	Estimate the overall concentration of Chlorophyll, AFDM,  Index, or the density of epilithic individuals, per sampling site and per type of habitat within the designated site (stone size class);
+
+
[…]
