### 2019-10-03 #345 @melvidoni
Hello @SD1060. Thanks for the repo link, but it still has no code inside of it. Based on the information you provided, it seems a like a methods / "methods workflow" package, which would be out of scope. We do hope the best for you, and that you can find an appropriate venue for it. Good luck!
