### 2019-10-04 #345 @SD1060
I have a question:
SDesti is a statistical analysis package, its objective is to facilitate the data analysis of two of the most widely used bio-indicators for the health of aquatic systems and grant consistency in the results and facilitate repetition the integration of data and results from different studies in decision making and management.
Would the above count as "reproducibility: Tools that facilitate reproducible research."?
Thank you,
Diogo
