### 2019-10-06 #345 @melvidoni
Hello @SD1060, as EiC I can take your argument to the rest of the editors, but we need the repo of your package to be completed with the code, if we want to reconsider this. Can you please do that, and let me know?
