### 2019-10-06 #345 @melvidoni
Hello @SD1060. Our _reproducibility_ category actually refers to a few distinct subtopics and that we realise that it can sound like a catch-all, which is why we just updated the guide to separate these out for more clarity ([check here](https://devguide.ropensci.org/policies.html#aims-and-scope)).

There is also an upcoming blog post about that: https://github.com/ropensci/roweb2/pull/521/files
