### 2019-10-09 #345 @SD1060
My problem is that I want to publish my program in Methods in Ecology and Evolution, they strongly recommend to use rOpenSci to validate our code before submitting a paper... I got rejected by the editor most probably because I didn't get my code reviewed, is there any other platform where I can get my program reviewed?
