### 2019-10-11 #345 @melvidoni
Hello @SD1060.

We are not sure where MEE "strongly recommends to use rOpenSci," as the partnership only works for packages that fit into both scopes. However, they are increasingly using our standards in the review.  Our guess is that the editor suggested that you use our standards to bring his package up to snuff to be reviewed.

We suggest submitting to JOSS, but similar journals will want the code hosted on GitHub or a similar platform, and at minimum pass checks for CRAN and have testing.  Our guide or the R packages book can provide guidance for how to do this.

If you encounter issues whilst trying to apply best practice from these sources then there are a few places where to ask specific questions, which you can find at: https://docs.r-hub.io/#places-to-ask-questions
